+++ src/js/shapes/iconsShape.js
@@ -1,12 +1,10 @@
 import PropTypes from 'prop-types';
 
 export default PropTypes.shape({
-    moveLeft: PropTypes.node.isRequired,
-    moveAllLeft: PropTypes.node.isRequired,
-    moveRight: PropTypes.node.isRequired,
-    moveAllRight: PropTypes.node.isRequired,
-
-    // Optional properties
+    moveLeft: PropTypes.node,
+    moveAllLeft: PropTypes.node,
+    moveRight: PropTypes.node,
+    moveAllRight: PropTypes.node,
     moveBottom: PropTypes.node,
     moveDown: PropTypes.node,
     moveUp: PropTypes.node,