+++ src/js/shapes/languageShape.js
@@ -1,15 +1,13 @@
 import PropTypes from 'prop-types';
 
 export default PropTypes.shape({
-    moveLeft: PropTypes.string.isRequired,
-    moveAllLeft: PropTypes.string.isRequired,
-    moveRight: PropTypes.string.isRequired,
-    moveAllRight: PropTypes.string.isRequired,
-
-    // Optional properties
     availableFilterHeader: PropTypes.string,
     availableHeader: PropTypes.string,
     filterPlaceholder: PropTypes.string,
+    moveLeft: PropTypes.string,
+    moveAllLeft: PropTypes.string,
+    moveRight: PropTypes.string,
+    moveAllRight: PropTypes.string,
     moveDown: PropTypes.string,
     moveUp: PropTypes.string,
     noAvailableOptions: PropTypes.string,