+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 
 #### Properties
 
+* `icons`: supplied value will now merge missing keys with the default
 * `id`: no longer defaults to a random UUID when null and no longer used for some child components
+* `lang`: supplied value will now merge missing keys with the default
 * `options`: no longer has PropTypes validation for `label` and `value`
 * `simpleValue`: no longer available; use `ObjectValueWrapper` instead and pass in your regular properties: